+++ CHANGELOG.md
@@ -5,6 +5,19 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 [Keep a CHANGELOG](http://keepachangelog.com).
 
 
+## Unreleased
+
+### Added 
+
+- new Document configuration `label_type` which can be set in the preamble
+  with the SKOS property to be used for producing the label statements;
+  for now only `:prefLabel` (the default) and `notation` are allowed
+
+
+[Compare v0.1.0...HEAD](https://github.com/marcelotto/skout/compare/v0.1....HEAD)
+
+
+
 ## v0.1.0 - 2019-09-09
 
 Initial release